adding code and explanation
The code explained line by line:

1. <?php - Indicates the beginning of the PHP block
2. $validacoes = []; - Creates an empty array to store validation messages
3. if (count($_POST) > 0) { - Checks if there is data sent through the form
4. if ($_POST['nome'] === "") { - Checks if the "nome" field is empty
5. $validacoes[] = "Por favor, preencha o nome do usuário !"; - Adds a validation message to the array
6. if ($_POST['email'] === "") { - Checks if the "email" field is empty
7. $validacoes[] = "Por favor, preencha o e-mail do usuário !"; - Adds a validation message to the array
8. if ($_POST['senha'] != $_POST['confirmacao']) { - Checks if the password and confirmation password fields are equal
9. $validacoes[] = "Senhas diferentes, verifique !"; - Adds a validation message to the array
10. } - Closes the condition block of item 8
11. } - Closes the condition block of item 3
12. ?> - Indicates the end of the PHP block

Starting from line 14, HTML code is written. On line 19, there is a condition that checks if the validation message array has any messages. If there are, a list of validation messages is displayed.

The rest of the code is a simple HTML form with fields for "nome", "email", "senha", and "confirmacao de senha" (confirmation of password). When the form is submitted, validations are performed, and if there are any validation messages, they are displayed on the page. Otherwise, the form is submitted.